requirements.txt
Jinja2==3.0.1
click==8.0.1
requests==2.26.0
validators==0.18.2
minio==7.1.0
rich==10.6.0
PyYAML==5.4.1
coverage==5.5
docker==5.0.0
